Imroved my code and updated hour log
Last week was not very productive for my assignment. During the week I work at my part time job. This is the company where I had my internship at the time of my MBO course. It was my intention to continue with the assignment on the Friday after work but instead I got sick. It was not Corona but a flu of some kind. Luckily for me, I have improved a lot in the last few days. During the last 2 days I have made some improvements to my application. I reflected on the way I handle errors. This could be done without the try catch method because this was a very cheap answer. To improve this I anticipated what parts of the code could give an error and tried to minimize the result. After the optimizations I am planning on implementing the bonus features. I have to keep in mind that it has already been a week since I’ve started on the assignment and that I should not take too much time.